LayoutTests/http/tests/security/contentSecurityPolicy/function-constructor-blocked-expected.txt

CONSOLE MESSAGE: line 12: EvalError: Function constructor is disabled
